:building_construction: Rangement du projet
- Ajout d'un gitignore
- Suppression des fichiers '/target'
- Désimbrication du projet
- Renommage du projet en 'fr.insa.christine.devis'
