--- mavenproject3/src/main/java/frchristineinsa/mavenproject3/Appartement.java
@@ -1,43 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package frchristineinsa.mavenproject3;
-
-import java.util.List;
-
-/**
- *
- * @author cesse
- */
-public class Appartement {
-    private int idAppartement ;
-    private int idNiveauAppartement ;
-    private List<Piece> listePieces ;
-
-    public Appartement(int idAppartement, int idNiveauAppartement, List<Piece> listePieces) {
-        this.idAppartement = idAppartement;
-        this.idNiveauAppartement = idNiveauAppartement;
-        this.listePieces = listePieces;
-    }
-
-    public int getIdAppartement() {
-        return idAppartement;
-    }
-
-    public int getIdNiveauAppartement() {
-        return idNiveauAppartement;
-    }
-
-    public List<Piece> getListePieces() {
-        return listePieces;
-    }
-
-    @Override
-    public String toString() {
-        return "Appartement{" + "idAppartement=" + idAppartement + ", idNiveauAppartement=" + idNiveauAppartement + ", listePieces=" + listePieces + '}';
-    }
-    
-    // méthode surface appart + montant revetement
-    
-}
