--- mavenproject3/src/main/java/frchristineinsa/mavenproject3/Fenetre.java
@@ -1,17 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package frchristineinsa.mavenproject3;
-
-/**
- *
- * @author cesse
- */
-public class Fenetre extends Ouverture {
-    
-    public Fenetre(int idOuverture, double dimX, double dimY) {
-        super(idOuverture, dimX, dimY);
-    }
-    
-}
