--- mavenproject3/src/main/java/frchristineinsa/mavenproject3/Immeuble.java
@@ -1,19 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package frchristineinsa.mavenproject3;
-
-import java.util.List;
-
-/**
- *
- * @author cesse
- */
-public class Immeuble extends Batiment{
-    
-    public Immeuble(String idBatiment, List<Niveau> listeNiveaux) {
-        super(idBatiment, listeNiveaux);
-    }
-    
-}
