--- mavenproject3/src/main/java/frchristineinsa/mavenproject3/Maison.java
@@ -1,22 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package frchristineinsa.mavenproject3;
-
-import java.util.List;
-
-/**
- *
- * @author cesse
- */
-public class Maison extends Batiment {
-
-    public Maison(String idBatiment, List<Niveau> listeNiveaux) { // on admet nb niveaux = 1 ou pas? sinon aucune différence avec immeuble
-        super(idBatiment, listeNiveaux);
-    }
-
-    
-    
-    
-}
