--- mavenproject3/src/main/java/frchristineinsa/mavenproject3/Porte.java
@@ -1,18 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package frchristineinsa.mavenproject3;
-
-/**
- *
- * @author cesse
- */
-public class Porte extends Ouverture {
-    
-    public Porte(int idOuverture, double dimX, double dimY) {
-        super(idOuverture, dimX, dimY);
-    }
-    
-    
-}
