--- mavenproject3/src/main/java/frchristineinsa/mavenproject3/SystemInfo.java
@@ -1,13 +1,0 @@
-package frchristineinsa.mavenproject3;
-
-public class SystemInfo {
-
-    public static String javaVersion() {
-        return System.getProperty("java.version");
-    }
-
-    public static String javafxVersion() {
-        return System.getProperty("javafx.version");
-    }
-
-}
