--- mavenproject3/src/main/java/module-info.java
@@ -1,4 +1,0 @@
-module frchristineinsa.mavenproject3 {
-    requires javafx.controls;
-    exports frchristineinsa.mavenproject3;
-}
