+++ src/main/java/fr/insa/christine/devis/Fenetre.java
@@ -1,0 +1,17 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package fr.insa.christine.devis;
+
+/**
+ *
+ * @author cesse
+ */
+public class Fenetre extends Ouverture {
+    
+    public Fenetre(int idOuverture, double dimX, double dimY) {
+        super(idOuverture, dimX, dimY);
+    }
+    
+}
