+++ src/main/java/fr/insa/christine/devis/Immeuble.java
@@ -1,0 +1,19 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package fr.insa.christine.devis;
+
+import java.util.List;
+
+/**
+ *
+ * @author cesse
+ */
+public class Immeuble extends Batiment{
+    
+    public Immeuble(String idBatiment, List<Niveau> listeNiveaux) {
+        super(idBatiment, listeNiveaux);
+    }
+    
+}
