+++ src/main/java/fr/insa/christine/devis/Maison.java
@@ -1,0 +1,22 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package fr.insa.christine.devis;
+
+import java.util.List;
+
+/**
+ *
+ * @author cesse
+ */
+public class Maison extends Batiment {
+
+    public Maison(String idBatiment, List<Niveau> listeNiveaux) { // on admet nb niveaux = 1 ou pas? sinon aucune différence avec immeuble
+        super(idBatiment, listeNiveaux);
+    }
+
+    
+    
+    
+}
