+++ src/main/java/fr/insa/christine/devis/SystemInfo.java
@@ -1,0 +1,13 @@
+package fr.insa.christine.devis;
+
+public class SystemInfo {
+
+    public static String javaVersion() {
+        return System.getProperty("java.version");
+    }
+
+    public static String javafxVersion() {
+        return System.getProperty("javafx.version");
+    }
+
+}
