+++ src/main/java/module-info.java
@@ -1,0 +1,4 @@
+module fr.insa.christine.devis {
+    requires javafx.controls;
+    exports fr.insa.christine.devis;
+}
